requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
rasterio>=1.3.0
pillow>=10.0.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
scipy>=1.11.0

# FAISS for vector search
faiss-cpu>=1.7.4

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
aiofiles>=23.0.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Visualization & Metrics
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
pandas>=2.0.0
shapely>=2.0.0
